[#182] Remove StablecoinImpl m capability
Problem: no longer required, we can call stablecoin client from network
scenarios directly.

Solution: remove the capability, rework helper functions to run in
MonadNetwork.
