+++ haskell/src/Stablecoin/Client/Cleveland.hs
@@ -3,7 +3,6 @@
 
 module Stablecoin.Client.Cleveland
   ( module Stablecoin.Client.Cleveland.Caps
-  , StablecoinImpl
   , Alias
   , AddressAndAlias(..)
   ) where
@@ -12,4 +11,3 @@ import Morley.Tezos.Address.Alias (Alias)
 
 import Stablecoin.Client (AddressAndAlias(..))
 import Stablecoin.Client.Cleveland.Caps
-import Stablecoin.Client.Cleveland.StablecoinImpl (StablecoinImpl)
